+++ src/RecoveryServices/RecoveryServices.SiteRecovery/ReplicationProtectedItem/NewAzureRmRecoveryServicesAsrReplicationProtectedItem.cs
@@ -299,6 +299,14 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         public string RecoveryProximityPlacementGroupId { get; set; }
 
         /// <summary>
+        /// Gets or sets the resource ID of virtual machine scale set to failover this virtual machine to.
+        /// </summary>
+        [Parameter(ParameterSetName = ASRParameterSets.AzureToAzure, HelpMessage = "Specify the virtual machine scale set Id to be used by the failover Vm in target recovery region.")]
+        [Parameter(ParameterSetName = ASRParameterSets.AzureToAzureWithoutDiskDetails, HelpMessage = "Specify the virtual machine scale set Id to be used by the failover Vm in target recovery region.")]
+        [ValidateNotNullOrEmpty]
+        public string RecoveryVirtualMachineScaleSetId { get; set; }
+
+        /// <summary>
         /// Gets or sets ID of the AvailabilitySet to recover the machine to in the event of a failover.
         /// </summary>
         [Parameter(ParameterSetName = ASRParameterSets.AzureToAzure)]
@@ -794,7 +802,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                 RecoveryAzureNetworkId = this.RecoveryAzureNetworkId,
                 RecoverySubnetName = this.RecoveryAzureSubnetName,
                 RecoveryAvailabilityZone = this.RecoveryAvailabilityZone,
-                RecoveryProximityPlacementGroupId = this.RecoveryProximityPlacementGroupId
+                RecoveryProximityPlacementGroupId = this.RecoveryProximityPlacementGroupId,
+                RecoveryVirtualMachineScaleSetId = this.RecoveryVirtualMachineScaleSetId
             };
 
             if (!string.IsNullOrEmpty(this.ReplicationGroupName))